+++ proxy/php/servers.php
@@ -1,0 +1,8 @@
+<?php
+
+return $servers = [
+    "node1" => "127.0.0.1:3085",  // Change to your real server address
+    "node2" => "127.0.0.2:3085",  // Change to your real server address
+    "node3" => "127.0.0.3:3085"   // Change to your real server address
+];
+