+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+### 1.0.1
++ [x] Added php proxy server
++ [x] Added uptime values from sidecar (parsed from http://uptime.minaprotocol.com)
+
 ### 1.0.0
 + [x] First release